fix null pointer exception in rsql target field validation test
