+++ hawkbit-repository/hawkbit-repository-jpa/src/test/java/org/eclipse/hawkbit/repository/jpa/rsql/RSQLTargetFieldTest.java
@@ -9,6 +9,7 @@
 package org.eclipse.hawkbit.repository.jpa.rsql;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import java.util.Arrays;
@@ -232,6 +233,32 @@ public class RSQLTargetFieldTest extends AbstractJpaIntegrationTest {
                 0);
         assertRSQLQuery("(" + TargetFields.TAG.name() + "!=TAG1" + OR + TargetFields.TAG.name() + "!=TAG2)" + AND
                 + TargetFields.CONTROLLERID.name() + "!=targetId1235", 4);
+    }
+
+    @Test
+    @Description("Testing allowed RSQL keys based on TargetFields definition")
+    public void rsqlValidTargetFields() {
+        final String rsql1 = "ID == '0123' and NAME == abcd and DESCRIPTION == absd"
+                + " and CREATEDAT =lt= 0123 and LASTMODIFIEDAT =gt= 0123"
+                + " and CONTROLLERID == 0123 and UPDATESTATUS == PENDING"
+                + " and IPADDRESS == 0123 and LASTCONTROLLERREQUESTAT == 0123" + " and tag == beta";
+
+        RSQLUtility.validateRsqlFor(rsql1, TargetFields.class);
+
+        final String rsql2 = "ASSIGNEDDS.name == abcd and ASSIGNEDDS.version == 0123"
+                + " and INSTALLEDDS.name == abcd and INSTALLEDDS.version == 0123";
+        RSQLUtility.validateRsqlFor(rsql2, TargetFields.class);
+
+        final String rsql3 = "ATTRIBUTE.subkey1 == test and ATTRIBUTE.subkey2 == test"
+                + " and METADATA.metakey1 == abcd and METADATA.metavalue2 == asdfg";
+        RSQLUtility.validateRsqlFor(rsql3, TargetFields.class);
+
+        final String rsql4 = "CREATEDAT =lt= ${NOW_TS} and LASTMODIFIEDAT =ge= ${OVERDUE_TS}";
+        RSQLUtility.validateRsqlFor(rsql4, TargetFields.class);
+
+        final String rsql5 = "wrongfield == abcd";
+        assertThatExceptionOfType(RSQLParameterUnsupportedFieldException.class)
+                .isThrownBy(() -> RSQLUtility.validateRsqlFor(rsql5, TargetFields.class));
     }
 
     private void assertRSQLQuery(final String rsqlParam, final long expcetedTargets) {
